Allow custom parameter resolution for table tests
